fix: #83 리뷰 반영
