+++ src/components/auth/Form/AdditionalInfoForm/index.tsx
@@ -5,7 +5,7 @@
 import PrimaryButton from '@/components/@common/Button/PrimaryButton';
 import CheckBox from '@/components/@common/Checkbox';
 import Radio from '@/components/@common/Radio';
-import { EI, FT, JP, NS } from '@/constants/mbti';
+import { EI, FT, JP, NS } from '@/constants/user/mbti';
 import { useFetchCategory } from '@/hooks/auth/useFetchCategory';
 import { useGeoLocation } from '@/hooks/auth/useGeoLocation';
 import signUpStepAtom from '@/recoil/auth/signUp/atom';
