+++ src/components/auth/Form/BasicInfoForm/index.tsx
@@ -6,31 +6,35 @@
 import PrimaryButton from '@/components/@common/Button/PrimaryButton';
 import Dropdown from '@/components/@common/Dropdown';
 import Radio from '@/components/@common/Radio';
-import { GENDER } from '@/constants/gender';
 import { AUTH_ERROR_MSG } from '@/constants/message';
-import { BIRTH_MONTH } from '@/constants/month';
 import { AUTH_PATTERN } from '@/constants/pattern';
-import { BIRTH_YEAR } from '@/constants/year';
 import { useCheckAvailability } from '@/hooks/auth/useCheckAvailability';
 import signUpStepAtom from '@/recoil/auth/signUp/atom';
 import { getDaysInMonth } from '@/utils/date';
 import { moveToStep } from '@/utils/step/moveSteps';
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 import { useSetRecoilState } from 'recoil';
 import ImagePicker from '../../ImagePicker';
 import InputField from '../../InputField';
 import * as Styled from './BasicInfoForm.styled';
+import { BIRTH_YEAR } from '@/constants/user/year';
+import { BIRTH_MONTH } from '@/constants/user/month';
+import { GENDER } from '@/constants/user/gender';
 
 const BasicInfoForm = () => {
+  const daySelectRef = useRef<any>(null);
+
   const setSignUpStep = useSetRecoilState(signUpStepAtom);
 
   const [days, setDays] = useState([{ value: '', label: '' }]);
 
+  const [prevYear, setPrevYear] = useState(null);
+  const [prevMonth, setPrevMonth] = useState(null);
+
   const {
     control,
     getValues,
-    setValue,
     formState: { isValid },
   } = useFormContext();
 
@@ -51,14 +55,24 @@ const BasicInfoForm = () => {
   };
 
   const handleSelectBirthYM = () => {
-    // 출생년도나 생월의 값이 변경되면 생일의 값을 1로 초기화
-    setValue('day', '01');
-
     // 사용자가 선택한 출생년도
     const year = getValues('year');
 
     // 사용자가 선택한 생월
     const month = getValues('month');
+
+    // 사용자가 선택한 생일
+    const day = getValues('day');
+
+    // 출생년도나 생월의 값이 변경되면 생일의 값을 초기화
+    if (day && (prevYear !== year || prevMonth !== month)) {
+      if (daySelectRef.current) {
+        daySelectRef.current.clearValue();
+      }
+    }
+
+    setPrevYear(year);
+    setPrevMonth(month);
 
     // 날짜 select에 날짜 세팅
     setDays(getDaysInMonth(year, month));
@@ -105,7 +119,7 @@ const BasicInfoForm = () => {
               render={({ field: { ref, value, onChange } }) => (
                 <Dropdown
                   name="year"
-                  label="나이(출생년도)"
+                  label="나이"
                   options={BIRTH_YEAR}
                   placeholder="출생년도를 선택해주세요."
                   isRequired={true}
@@ -150,19 +164,19 @@ const BasicInfoForm = () => {
               rules={{
                 required: {
                   value: true,
-                  message: '일을 선택해주세요.',
+                  message: '일자를 선택해주세요',
                 },
               }}
-              render={({ field: { ref, value, onChange } }) => (
+              render={({ field: { value, onChange } }) => (
                 <Dropdown
                   name="day"
                   width="156px"
                   options={days}
                   placeholder="일"
-                  ref={ref}
+                  ref={daySelectRef}
                   value={days.find((option) => option.value === value)}
                   onChange={(...args) => {
-                    onChange(args[0].value);
+                    onChange(args[0] ? args[0].value : null);
                   }}
                 />
               )}
