+++ src/constants/user/mbti/index.ts
